Move AddPlanDisabledObjModel into plans page types file
