--- src/app/home/AddPlanDisabledObjModel.ts
@@ -1,4 +1,0 @@
-export interface IAddPlanDisabledObjModel {
-  isAddPlanDisabled: boolean;
-  disabledText: string;
-}
